made all requested changes
