+++ client/website/src/pages/JoinUs.js
@@ -1,16 +1,31 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import ResourcesHeader from "../components/ResourcesHeader";
-import HeaderImage from "../media/JoinUs_Header.png";
-import MobileHeaderImage from "../media/JoinUs_Header_Mobile.svg";
 
 export default function JoinUs() {
+    const [isMobile, setIsMobile] = useState(false);
+
+    // modifies isMobile state when window resizes
+    useEffect(() => {
+        function handleResize() {
+            setIsMobile(window.innerWidth <= 600);
+        }
+
+        // event listener for resize
+        window.addEventListener("resize", handleResize);
+        handleResize();
+
+        // Removes event listener on cleanup
+        return () => window.removeEventListener("resize", handleResize);
+    }, []);
+
     return (
         <div>
             <ResourcesHeader
-                image={HeaderImage}
-                mobileImage={MobileHeaderImage}
+                image="https://s3-alpha-sig.figma.com/img/2632/7b67/21741a7000235fce33b207b71e8da1cd?Expires=1622419200&Signature=NaKrOXhTogZlNaWaE-kcQHGyv7cZL745P19PBIu3a~8UWgMEh0wqj3bAmfbhisl4j91lMJQei5kgDCFq4KIUeqRX39BNJxg7qt8CKrpx1Hw-kdUaaTz-10Rqp6tw~vxCma9x19Ry~JIhDTRyAjyn2zN21NTYCNvF1Xpc2BgzzavRdZsW22cFD66fde8NG-Sqq85fT4QwvDxStiavUn0EqvAYPvYDxJ2uhghbnI6U07n5MmEI~nrUzV2IamBqywS-B5TMo7Qqxnajzf1H6WgNKrFRVFok5jNVy4J4HNXYJLp7lQFXZ1yzs9YD~bapJXBg1Yo9R6KTWsv58zxvtw2M6g__&Key-Pair-Id=APKAINTVSUGEWH5XD5UA"
                 title="Join Us"
                 text="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Maecenas facilisis condimentum massa, sit amet lacinia massa commodo sed. Praesent vehicula eget arcu ut laoreet. Sed porta, dui ut dapibus sodales, orci neque volutpat arcu, in efficitur sem tortor vel lectus."
+                height={isMobile ? "95vh" : "80.7vh"}
+                width="100%"
             />
             <p> This is the Join Us Page </p>
         </div>
